backend/requirements.txt
Django==5.0.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
psycopg[binary]>=3.2.10
python-decouple==3.8
gunicorn==21.2.0
